Update 36-snk-neo-geo-cd.md
I think this part is for neogeo, not for neogeo cd. Maybe it's the same for : 

"UniBIOS can be used but v4.0 can lead to missing or corrupted audio when used with save states. Earlier UniBIOS version are OK."

If it's also only for neogeo it can be removed too.